Add @remix-run/create script
This commit separates the `create-remix` script into 2 pieces: the
`create-remix` piece is responsible for creating the app directory and
setting up npm auth. Then it delegates to `@remix-run/create` to install
the templates and `package.json`.

This is really an elaborate workaround for the fact that npm caches
init scripts heavily and we need a way to run the latest version easily.