Fix hash change navigation
Previously we weren't updating the location, we just skipped everything! The point was to skip data loading, but totally forgot to update the location and other states.

Now it acts exactly like a normal pathname change navigation, except it skips the loaders (because it's just an internal document link). While we technically don't need to go into a pending state, UI abstractions like scroll/focus management will rely on the pending state to save off information (like the scroll position of the old page). My hunch is we'd rather have a fake pending state so that these abstractions don't have to special case the hash change, they can always plan on a pending state (reminds me of the "promises are always async to prevent zalgo" conversations years ago when promises were being added to the language. We may want to revisit this, but for now I think it's good.