chore(tests): match conventions in integration tests
We generally just want one fixture per test module since they are expensive. Having the fixture defined at the top of the file encourages us to keep adding to that fixture instead of making a bunch of them in each test.

Also moved the global-setup out of the root of integration/ since it just felt a little messy